operations/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.1
python-dotenv>=1.0.0
Pillow>=9.0.0
